feat(package/custom-container): ros suppert custom-container
