+++ src/bin/fun-package.js
@@ -17,6 +17,7 @@ program
   .option('-b, --oss-bucket <bucket>', 'The name of the oss bucket where Fun uploads local artifacts')
   .option('-o, --output-template-file <filename>', 'The output path of the packaged template file')
   .option('--use-nas', 'Automatically upload local resources to NAS.')
+  .option('--push-registry <pushRegistry>', 'Modify the image upload path')
   .parse(process.argv);
 
 if (program.args.length > 1) {
